requirements.txt
pika
pymongo>=3.7.2
tqdm
requests
flask
gevent
pyopenssl
